Auction - Add auth to the endpoints

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -49,12 +50,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
         {
             var auction = _mapper.Map<Auction>(createAuctionDto);
-            // TODO: Add current user as Seller
-            auction.Seller = "test";
+            
+            auction.Seller = User.Identity.Name;
 
             // Add to Auctions table
             _context.Auctions.Add(auction);
@@ -67,12 +69,13 @@ namespace AuctionService.Controllers
             var result = await _context.SaveChangesAsync() > 0;
 
             if (!result)
-                return BadRequest("Could not save auction to database");
+                return BadRequest();
 
             return CreatedAtAction(nameof(GetAuctionById), 
                 new {auction.Id}, _mapper.Map<AuctionDto>(auction));
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -81,8 +84,10 @@ namespace AuctionService.Controllers
             if (auction == null)
                 return NotFound();
 
-            // TODO: Check seller is current user
-
+            // Check seller is current user
+            if (auction.Seller != User.Identity.Name)
+                return Forbid();
+            
             // Update all fields
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -102,6 +107,7 @@ namespace AuctionService.Controllers
             return Ok();
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -110,7 +116,9 @@ namespace AuctionService.Controllers
             if (auction == null)
                 return NotFound();
             
-            // TODO: Check seller is current user
+            // Check seller is current user
+            if (auction.Seller != User.Identity.Name)
+                return Forbid("You are not allowed to delete this auction");
 
             _context.Auctions.Remove(auction);
